fix(sdk): keep Aleo runtime eager in Node
